fix: code review comments - build widgets indentation - section icon background color - creating sections folders for skills, bio, info
