+++ lib/profile/section_icons.dart
@@ -5,14 +5,14 @@ class ProfileSectionIcons extends StatelessWidget {
   const ProfileSectionIcons({Key key}) : super(key: key);
   
   @override
-    Widget build(BuildContext context) {
-      return Row(
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          ProfileIconButton(name: Icons.done, caption: "Skills"),
-          ProfileIconButton(name: Icons.subject, caption: "Bio"),
-          ProfileIconButton(name: Icons.person_outline, caption: "Info"),
-        ],
-      );
-    }
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        ProfileIconButton(name: Icons.done, caption: "Skills"),
+        ProfileIconButton(name: Icons.subject, caption: "Bio"),
+        ProfileIconButton(name: Icons.person_outline, caption: "Info"),
+      ],
+    );
+  }
 }
